% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleFun.R
\name{downscaleFun}
\alias{downscaleFun}
\title{Downscale MODIS scene}
\usage{
downscaleFun(
  MOD,
  ia,
  modname,
  aux = aux,
  template = template,
  writeRastertoOutdir = TRUE,
  outdir,
  add = NULL,
  outformat = ".grd"
)
}
\arguments{
\item{MOD}{MODIS scene in format raster, will be resampled to template resolution in case it is not already provided as such}

\item{ia}{incidence angle in format raster as generated by make_ia()}

\item{modname}{character, containing the first part of the MODIS scene name at least to derive whether the scene stems from Aqua or Terra satellite}

\item{aux}{auxiliary data in format raster stack as provided by package as aux.grd}

\item{template}{template in format raster as provided by package as template.grd}

\item{add}{character, for an additional suffix to the outfile name in order to mark e.g. smaller extens}

\item{outformat}{default is ".grd" but could also be ".tif"}
}
\value{
downscaled MODIS LST at 30m resolution
}
\description{
Downscale MODIS scene
}
\examples{
library(raster)
library(lubridate)
library(solrad)
library(downscaleLST.MDV)


exdate <- as.POSIXct("2018-11-12 13:50:00", tz="UTC")
ex2 <- raster::stack("inst/ex_2_all_layers.grd")

##########  load the model ###############################################
load(system.file("final_model_downscaling_LST_MDV.RData", package = "downscaleLST.MDV"))

##########  make ia and put into one stack  ###############################
ia_ex2 <- make_ia(M_date = exdate, outdir = here::here("ex_output"),
                     sl=ex2$slope, as=ex2$aspect, tmplt=ex2$template)
names(ia_ex2) <- "ia"
ex2_ia <- raster::stack(ex2, ia_ex2)


dsc_ex2 <- downscaleFun(MOD=ex2_ia$Mbilin,
                        aux=ex2_ia[[c( "dem","slope","aspect","soilraster","landcoverres")]],
                        ia=ex2_ia$ia, 
                        template=ex2_ia$template, 
                        modname="MYD11_L2.A2018316.1350.006",
                        outdir = "ex_output/",
                        writeRastertoOutdir = TRUE,
                        add="_ex2",
                        outformat=".grd")

comp_ex2 <- raster::stack(ex2$Mbilin, ex2$Mngb, dsc_ex2, ex2$L)
raster::writeRaster(comp_ex2, "inst/comp_ex2.grd",overwrite=T)

}
