% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ia.R
\name{make_ia}
\alias{make_ia}
\title{Calculate incidence angle for MODIS scene to be downscaled}
\usage{
make_ia(
  M_date,
  outdir,
  tmpdir = tempdir(),
  sl = slope,
  as = aspect,
  tmplt = template,
  writeRastertoOutdir = TRUE
)
}
\arguments{
\item{M_date}{MODIS scene capturing time in format POSIXct, UTC, just as in the MODIS filename}

\item{outdir}{directory the incidence angle will be written to}

\item{tmpdir}{directory for writing temporary raster files}

\item{sl}{slope in degrees, full MDV file loadable via slope <- raster(system.file("30m_slopeMDV.tif", package = "downscaleLST.MDV"))}

\item{as}{aspect in degrees, full MDV file loadable via aspect <- raster(system.file("30m_aspectMDV.tif", package = "downscaleLST.MDV"))}

\item{template}{template file in format raster}
}
\value{
Incidence Angle raster for the time the MODIS scene to be downscaled was captured
}
\description{
Calculate incidence angle for MODIS scene to be downscaled
}
\examples{

exdate <- as.POSIXct("2018-11-12 13:50:00", tz="UTC")
full <- raster::stack("inst/full_size_grids_all_layers.grd")
ex2 <- raster::stack("inst/ex_2_all_layers.grd")

# for the example extent
ia <- make_ia(M_date = exdate, outdir = here::here("ex_day"), 
              sl=ex2$slope, as=ex2$aspect, tmplt=ex2$template)
# # for the full scene
# ia <- make_ia(M_date = exdate, outdir = here::here("ex_day"), 
#               sl=full$slope, as=full$aspect, tmplt=full$template)


}
