% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_median.R
\name{my_median}
\alias{my_median}
\title{My median}
\usage{
my_median(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector of Numeric values}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
Median of vector x
}
\description{
My median
}
\examples{
my_median(1:12)
}
