% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{make a figure for comparing MODIS to downscaled and reference scene}
\usage{
theme_map(y90 = TRUE, ts = 10)
}
\arguments{
\item{LST}{raster stack containing either MODIS and downscaled MODIS scene or the same plus a reference scene, preferably in this order, will be named by layer names}

\item{ex}{list containing extents that shall be plotted into the overview of the scene and displayed in a zoomed form below}

\item{c}{character string with R colours names to use for smaller extent headings and boxes in main figure, should be same length as list in ex}
}
\value{

}
\description{
this function was adapted from ggAOA by Ludwigm6 https://github.com/Ludwigm6/mlutils/blob/b8fab319f8054df6bb4595bd8241a2d50c7753e3/R/plot_aoa.R
}
\examples{
library(rasterVis)
library(sf)
library(viridis)
library(gridExtra)


comp_ex2 <- raster::stack("inst/comp_ex2.grd") # from example for downscaleFun

aoa <- raster::raster(list.files("inst/", pattern="AOA", full.names = T))

# this is the result of applying the 
aoa_ex2 <- raster::crop(aoa, comp_ex2)
aoa_ex2[aoa_ex2 == 1] = NA # set pixels that are within AOA to NA so  they don't overlap valid LST data

ex2 <- raster::stack(comp_ex2[[2:4]], aoa_ex2)
names(ex2) <- c("MODIS_LST", "downscaled_LST","Landsat_LST", "AOA")

# # stack original LST, downscaled LST and reference LST 
# M_dsc_L_stack <- raster::stack(full$Mngb, dsc_M, full$L, aoa)

# ex2 <- crop(M_dsc_L_stack, extent(dsc_ex2))

############### PLOTTING ###################################################
# 
# ### full scene ################
# dscplot <- make_dsc_figure(LST = M_dsc_L_stack$downscaled_LST,  LST_name = "LST",
#                  aoa=M_dsc_L_stack$AOA, ex=list(ex2, ex4), c=c("navy","chartreuse4"), 
#                  limlow = -40, limhigh = 0, exsize = 1.5, 
#                  limitx = c(350500, 500000), limity = c(-1400000, -1200000),legendposition = "none"
#                  )
# mplot <- make_dsc_figure(LST = M_dsc_L_stack$MODIS_LST,  LST_name = "MODIS LST",
#                  ex=list(ex2, ex4), c=c("navy","chartreuse4"), 
#                  limlow = -40, limhigh = 0, exsize = 1.5, 
#                  limitx = c(350500, 500050), limity = c(-1400050, -1200000),legendposition = "none"
#                )
# lplot <- make_dsc_figure(LST = M_dsc_L_stack$Landsat_LST,  LST_name = "Landsat LST",
#                ex=list(ex2, ex4), c=c("navy","chartreuse4"), 
#                limlow = -40, limhigh = 0, exsize = 1.5, 
#                limitx = c(350500, 500050), limity = c(-1400050, -1200000),legendposition = "none"
# )
# 
# plotlist <- list(mplot, dscplot, lplot)
# 
# margin = theme(plot.margin = unit(c(0.7,0.7,0.4,0.4), "cm"))
# 
# # png(paste0(figurepath, "full_example_scene_with_aoa.png"),
# #     units="cm", width=21, height=9, res=300)
# grid.arrange(
#   grobs = lapply(plotlist, "+", margin),
#   layout_matrix = rbind(c(1,2,3)))
# # dev.off()



### example 2 #################
library(ggplot2)
margin = ggplot2::theme(plot.margin = unit(c(0.4,0.4,0.4,0.4), "cm"))

dscplot <- make_dsc_figure(LST = ex2$downscaled_LST,  LST_name = "LST",
                 aoa=ex2$AOA, ex=list(ex2), c=c("navy"), 
                 limlow = -26, limhigh = -3, legendposition = "none",
                 exsize = 4)

mplot <- make_dsc_figure(LST = ex2$MODIS_LST,  LST_name = "MODIS LST",
               limlow = -26, limhigh = -3,ex=list(ex2), c=c("navy"), legendposition = "none",
               exsize = 4)

lplot <- make_dsc_figure(LST = ex2$Landsat_LST,  LST_name = "Landsat LST",
               limlow = -26, limhigh = -3,ex=list(ex2), c=c("navy"), legendposition = "none",
               exsize = 4)

plotlist <- list(mplot, dscplot, lplot)


gridExtra::grid.arrange(
  grobs = lapply(plotlist, "+", margin),
  layout_matrix = rbind(c(1,2,3)))
dev.off()

# with legend
dscplot_with_legend <- make_dsc_figure(LST = ex2$downscaled_LST,  LST_name = "LST",
                 aoa=ex2$AOA, ex=list(ex2), c=c("navy"), 
                 limlow = -26, limhigh = -3,
                 exsize = 2)


}
